Add firework setting and playing
