+++ Assets/Resources/UIPrefabs/FireworkSettingUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2152872fb8a4de94cb69523e3339cc82
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
