+++ Assets/Scenes/TestGym.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5f8fb6ce5928ddf48859533b54ec8264
+guid: edcbf99819a0c2f49b1722faf01296c5
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
