+++ Assets/Scripts/BagSystem/FireworkBagGrid.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0f864bee1e97b7f42b9735d7da25cf95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
