+++ Assets/Scripts/MouseController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0a0a6c30090817c4f8bd785aa90ce6b6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
