+++ Assets/Scripts/UISystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e79b5ac60e034fa4986c8d1833952fa5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
