+++ Assets/URP/URPAsset.asset
@@ -24,7 +24,7 @@ MonoBehaviour:
   m_OpaqueDownsampling: 0
   m_SupportsTerrainHoles: 0
   m_StoreActionsOptimization: 0
-  m_SupportsHDR: 0
+  m_SupportsHDR: 1
   m_MSAA: 1
   m_RenderScale: 1
   m_UpscalingFilter: 0
