+++ UserSettings/EditorUserSettings.asset
@@ -6,24 +6,12 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
-      value: 5a08575f5207595a0f5d59741173094444164f7d7d2a23317c7a4465bbe1646d
-      flags: 0
-    RecentlyUsedSceneGuid-1:
       value: 55500553030d5f5e0e0d5c7549225c4412154a2b292b20312c7d4c66e1e1623c
       flags: 0
-    RecentlyUsedSceneGuid-2:
-      value: 5606075451500d030e5e5c7016225b444e4f492978717e65297f1b63b2b3673e
-      flags: 0
-    RecentlyUsedSceneGuid-3:
-      value: 5203525501515a030e58557447705d44464f41797d7c24627e704d6be1e46339
-      flags: 0
-    RecentlyUsedSceneGuid-4:
-      value: 5101030357575e5958570e7b16700744124f4b7e2f7075617f794965b7b86368
-      flags: 0
-    RecentlyUsedSceneGuid-5:
+    RecentlyUsedSceneGuid-1:
       value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
       flags: 0
-    RecentlyUsedSceneGuid-6:
+    RecentlyUsedSceneGuid-2:
       value: 5609020403570d0d5c0c547045705c4442154879752e20327d791b31b3e2316e
       flags: 0
     RecentlyUsedScenePath-0:
@@ -37,7 +25,7 @@ EditorUserSettings:
   m_VCDebugCmd: 0
   m_VCDebugOut: 0
   m_SemanticMergeMode: 2
-  m_DesiredImportWorkerCount: 5
+  m_DesiredImportWorkerCount: 4
   m_StandbyImportWorkerCount: 2
   m_IdleImportWorkerShutdownDelay: 60000
   m_VCShowFailedCheckout: 1
